split dataset api into separate files and same for unit tests
