+++ api/datasets_api_test.go
@@ -431,1466 +431,6 @@ func TestGetDatasetAuditingErrors(t *testing.T) {
 	})
 }
 
-func TestGetEditionsReturnsOK(t *testing.T) {
-
-	t.Parallel()
-	Convey("A successful request to get edition returns 200 OK response", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return &models.EditionUpdateResults{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusOK)
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionSuccessful, auditParams)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetEditionsAuditingError(t *testing.T) {
-	t.Parallel()
-	Convey("given auditing get editions attempted action returns an error then a 500 response is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return &models.EditionUpdateResults{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			return errors.New("get editions action attempted audit event error")
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 1)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("given auditing get editions action successful returns an error then a 500 response is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return &models.EditionUpdateResults{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionsAction && result == actionSuccessful {
-				return errors.New("audit error")
-			}
-			return nil
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionSuccessful, auditParams)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When the dataset does not exist and auditing the action result fails then return status 500", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		r.Header.Add("internal-token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionsAction && result == actionUnsuccessful {
-				return errors.New(auditError)
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionUnsuccessful, auditParams)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When no published editions exist against a published dataset and auditing unsuccessful errors return status 500", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return nil, errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionsAction && result == actionUnsuccessful {
-				return errors.New(auditError)
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionUnsuccessful, auditParams)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetEditionsReturnsError(t *testing.T) {
-	t.Parallel()
-	Convey("When the api cannot connect to datastore return an internal server error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errInternal
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "internal error\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionUnsuccessful, auditParams)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset does not exist return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		r.Header.Add("internal-token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Dataset not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionUnsuccessful, auditParams)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When no editions exist against an existing dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		r.Header.Add("internal-token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return nil, errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionsAction, actionAttempted, auditParams)
-		verifyAuditRecordCalls(recCalls[1], getEditionsAction, actionUnsuccessful, auditParams)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When no published editions exist against a published dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionsFunc: func(id string, state string) (*models.EditionUpdateResults, error) {
-				return nil, errs.ErrEditionNotFound
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor())
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetEditionReturnsOK(t *testing.T) {
-	t.Parallel()
-	Convey("A successful request to get edition returns 200 OK response", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionFunc: func(id string, editionID string, state string) (*models.EditionUpdate, error) {
-				return &models.EditionUpdate{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetEditionReturnsError(t *testing.T) {
-	t.Parallel()
-	Convey("When the api cannot connect to datastore return an internal server error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errInternal
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "internal error\n")
-
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset does not exist return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		r.Header.Add("internal-token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Dataset not found\n")
-
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When edition does not exist for a dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		r.Header.Add("internal-token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionFunc: func(id string, editionID string, state string) (*models.EditionUpdate, error) {
-				return nil, errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When edition is not published for a dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionFunc: func(id string, editionID string, state string) (*models.EditionUpdate, error) {
-				return nil, errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetEditionAuditErrors(t *testing.T) {
-	t.Parallel()
-	Convey("when auditing get edition attempted action errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			return errors.New("auditing error")
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 1)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("when check dataset exists errors and auditing action unsuccessful errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return errors.New("check dataset error")
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionAction && result == actionUnsuccessful {
-				return errors.New("auditing error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("when check edition exists errors and auditing action unsuccessful errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return nil
-			},
-			GetEditionFunc: func(ID, editionID, state string) (*models.EditionUpdate, error) {
-				return nil, errors.New("get edition error")
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionAction && result == actionUnsuccessful {
-				return errors.New("auditing error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("when get edition audit even successful errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			GetEditionFunc: func(id string, editionID string, state string) (*models.EditionUpdate, error) {
-				return &models.EditionUpdate{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getEditionAction && result == actionSuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getEditionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getEditionAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetEditionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionsReturnsOK(t *testing.T) {
-	t.Parallel()
-	Convey("A successful request to get version returns 200 OK response", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID, editionID, state string) (*models.VersionResults, error) {
-				return &models.VersionResults{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionsReturnsError(t *testing.T) {
-	t.Parallel()
-	p := common.Params{"dataset_id": "123-456", "edition": "678"}
-
-	Convey("When the api cannot connect to datastore return an internal server error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errInternal
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "internal error\n")
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset does not exist return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Dataset not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the edition of a dataset does not exist return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When version does not exist for an edition of a dataset returns status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID, editionID, state string) (*models.VersionResults, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Version not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When version is not published against an edition of a dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID, editionID, state string) (*models.VersionResults, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Version not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When a published version has an incorrect state for an edition of a dataset return an internal error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-
-		version := models.Version{State: "gobbly-gook"}
-		items := []models.Version{version}
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID, editionID, state string) (*models.VersionResults, error) {
-				return &models.VersionResults{Items: items}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "Incorrect resource state\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionsAuditError(t *testing.T) {
-	t.Parallel()
-	p := common.Params{"dataset_id": "123-456", "edition": "678"}
-	err := errors.New("error")
-
-	Convey("when auditing get versions attempted action errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			return err
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 1)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("when auditing check dataset exists error returns an error then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return err
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionsAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("when auditing check edition exists error returns an error then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(ID string, editionID string, state string) error {
-				return err
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionsAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("when auditing get versions error returns an error then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(ID string, editionID string, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID string, editionID string, state string) (*models.VersionResults, error) {
-				return nil, err
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionsAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("when auditing invalid state returns an error then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(ID string, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(ID string, editionID string, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID string, editionID string, state string) (*models.VersionResults, error) {
-				return &models.VersionResults{
-					Items: []models.Version{{State: "not valid"}},
-				}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionsAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionUnsuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("when auditing get versions successful event errors then an 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionsFunc: func(datasetID, editionID, state string) (*models.VersionResults, error) {
-				return &models.VersionResults{}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionsAction && result == actionSuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		p := common.Params{"dataset_id": "123-456", "edition": "678"}
-		recCalls := auditMock.RecordCalls()
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionsAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionsAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionsCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionReturnsOK(t *testing.T) {
-	t.Parallel()
-	Convey("A successful request to get version returns 200 OK response", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return &models.Version{
-					State: models.EditionConfirmedState,
-					Links: &models.VersionLinks{
-						Self: &models.LinkObject{},
-						Version: &models.LinkObject{
-							HRef: "href",
-						},
-					},
-				}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusOK)
-		p := common.Params{"dataset_id": "123-456", "edition": "678", "version": "1"}
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionReturnsError(t *testing.T) {
-	p := common.Params{"dataset_id": "123-456", "edition": "678", "version": "1"}
-	t.Parallel()
-	Convey("When the api cannot connect to datastore return an internal server error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errInternal
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "internal error\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When the dataset does not exist for return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Dataset not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the edition of a dataset does not exist return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Edition not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When version does not exist for an edition of a dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Version not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When version is not published for an edition of a dataset return status not found", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldResemble, "Version not found\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When an unpublished version has an incorrect state for an edition of a dataset return an internal error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return &models.Version{
-					State: "gobbly-gook",
-					Links: &models.VersionLinks{
-						Self: &models.LinkObject{},
-						Version: &models.LinkObject{
-							HRef: "href",
-						},
-					},
-				}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, "Incorrect resource state\n")
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetVersionAuditErrors(t *testing.T) {
-	p := common.Params{"dataset_id": "123-456", "edition": "678", "version": "1"}
-	t.Parallel()
-	Convey("When auditing get version action attempted errors then a 500 status is returned", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errInternal
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionAttempted {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 1)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset does not exist and audit errors then return a 500 status", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return errs.ErrDatasetNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldResemble, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the edition does not exist for a dataset and auditing errors then a 500 status", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return errs.ErrEditionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When version does not exist for an edition of a dataset and auditing errors then a 500 status", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When version does not exist for an edition of a dataset and auditing errors then a 500 status", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-		r.Header.Add("internal_token", "coffee")
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return &models.Version{
-					State: "indifferent",
-					Links: &models.VersionLinks{
-						Self:    &models.LinkObject{HRef: "self"},
-						Version: &models.LinkObject{HRef: "version"},
-					},
-				}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionUnsuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionUnsuccessful, p)
-
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("when auditing a successful request to get a version errors then return a 500 status", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123-456/editions/678/versions/1", nil)
-
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			CheckDatasetExistsFunc: func(datasetID, state string) error {
-				return nil
-			},
-			CheckEditionExistsFunc: func(datasetID, editionID, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, editionID, version, state string) (*models.Version, error) {
-				return &models.Version{
-					State: models.EditionConfirmedState,
-					Links: &models.VersionLinks{
-						Self: &models.LinkObject{},
-						Version: &models.LinkObject{
-							HRef: "href",
-						},
-					},
-				}, nil
-			},
-		}
-
-		auditMock := getMockAuditor()
-		auditMock.RecordFunc = func(ctx context.Context, action string, result string, params common.Params) error {
-			if action == getVersionAction && result == actionSuccessful {
-				return errors.New("error")
-			}
-			return nil
-		}
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, auditMock)
-
-		api.router.ServeHTTP(w, r)
-
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, internalServerErr)
-		recCalls := auditMock.RecordCalls()
-		So(len(recCalls), ShouldEqual, 2)
-		verifyAuditRecordCalls(recCalls[0], getVersionAction, actionAttempted, p)
-		verifyAuditRecordCalls(recCalls[1], getVersionAction, actionSuccessful, p)
-		So(len(mockedDataStore.CheckDatasetExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-}
-
 func TestPostDatasetsReturnsCreated(t *testing.T) {
 	t.Parallel()
 	Convey("A successful request to post dataset returns 200 OK response", t, func() {
